small corrections

+++ Source/common.h
@@ -22,7 +22,7 @@ typedef float sample_t;
 #define CLIP(x, min, max) (x > max ? max : x < min ? min : x)
 
 static inline float scale(float x, float min, float max, float newmin, float newmax, float curve) {
-    return std::powf((x - min) / (max - min), curve) * (newmax - newmin) + newmin;
+    return std::pow((x - min) / (max - min), curve) * (newmax - newmin) + newmin;
 }
